pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synndicate"
version = "2.0.0"
description = "Advanced Multi-Agent AI Orchestration System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "himokai", email = "himokai@example.com"},
]
keywords = ["ai", "agents", "orchestration", "rag", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    
    # HTTP and async
    "httpx>=0.27.2",
    "aiofiles>=24.1.0",
    "asyncio-throttle>=1.0.2",
    
    # AI/ML
    "openai>=1.51.0",
    "sentence-transformers>=3.1.1",
    "chromadb>=0.5.15",
    "faiss-cpu>=1.8.0",
    "tiktoken>=0.8.0",
    
    # Observability
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "structlog>=24.4.0",
    
    # Utilities
    "tenacity>=9.0.0",
    "rich>=13.8.1",
    "typer>=0.12.5",
    "psutil>=6.0.0",
    "docker>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "pre-commit>=4.0.1",
]

[project.scripts]
synndicate = "synndicate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/synndicate"]

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "TCH"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/synndicate --cov-report=term-missing --cov-report=html"
